Use layout as component
