+++ src/pages/index.js
@@ -1,13 +1,17 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
+import Layout from '../components/layout'
+
 const IndexPage = () => (
-  <div>
-    <h1>Hi people</h1>
-    <p>Welcome to your new Gatsby site.</p>
-    <p>Now go build something great.</p>
-    <Link to="/page-2/">Go to page 2</Link>
-  </div>
+  <Layout>
+    <>
+      <h1>Hi people</h1>
+      <p>Welcome to your new Gatsby site.</p>
+      <p>Now go build something great.</p>
+      <Link to="/page-2/">Go to page 2</Link>
+    </>
+  </Layout>
 )
 
 export default IndexPage
