+++ src/pages/page-2.js
@@ -1,12 +1,16 @@
 import React from 'react'
 import { Link } from 'gatsby'
 
+import Layout from '../components/layout'
+
 const SecondPage = () => (
-  <div>
-    <h1>Hi from the second page</h1>
-    <p>Welcome to page 2</p>
-    <Link to="/">Go back to the homepage</Link>
-  </div>
+  <Layout>
+    <>
+      <h1>Hi from the second page</h1>
+      <p>Welcome to page 2</p>
+      <Link to="/">Go back to the homepage</Link>
+    </>
+  </Layout>
 )
 
 export default SecondPage
